refactor: modular, composable lightning service
- export all gRPC message types
- uses small functions rather than building an object literal
- allows consumers to perform any API call
